refactor existing rustls usage into its own module
